refactor: separate qr transports from modal qr display
Move send and chasquiSend out of transports/qr/index into a new file transports/qr.js
Create new folder transport/ui to exclusively hold UI elements and functions for displaying QR modals and notifications
Separate svg elements defined in transport/ui/assets.js into individual SVG files in transport/ui/assets
