+++ src/transport/ui/assets/index.js
@@ -1,0 +1,11 @@
+import logo from './logo.svg'
+import logoWithBG from './logo-bg.svg'
+import appleApp from './app-store-badge.svg'
+import androidApp from './google-play-badge.svg'
+import close from './close-x.svg'
+
+const SVG = {
+	logo, logoWithBG, appleApp, androidApp, close
+}
+
+export default SVG
